feat: log settings and constants updated
