+++ src/main/interfaces/MigrationInterface.ts
@@ -1,0 +1,10 @@
+export interface Migration {
+  id: string;
+  checksum: string;
+  finished_at: string;
+  migration_name: string;
+  logs: string;
+  rolled_back_at: string;
+  started_at: string;
+  applied_steps_count: string;
+}
