+++ src/main/utils/runPrismaCommand.ts
@@ -2,7 +2,7 @@
 import path from 'path';
 import log from 'electron-log';
 import { fork } from 'child_process';
-import { mePath, qePath } from './defaultConstants';
+import { mePath, qePath } from '../../common/utils/globalConstants';
 
 export default async function runPrismaCommand({
   command,
