+++ src/renderer/App.tsx
@@ -1,5 +1,6 @@
 import { Switch, Route, useLocation } from 'react-router-dom';
 import { AnimatePresence } from 'framer-motion';
+import log from 'electron-log';
 
 // views / components
 import EnvironmentView from './views/EnvironmentView';
@@ -15,8 +16,16 @@ import './assets/styles/utilities.scss';
 // contexts
 import { EnvironmentListContextProvider } from './contexts/EnvironmentListContext';
 import { NotificationsContextProvider } from './contexts/NotificationsContext';
+import {
+  isDevelopment,
+  logStringFormat,
+} from '../common/utils/globalConstants';
 
 export default function App() {
+  log.transports.file.fileName = isDevelopment ? 'app.dev.log' : 'app.log';
+  log.transports.file.format = logStringFormat;
+
+  log.info('Renderer process started');
   // the useLocation hook is used to render a specific component per route
   const location = useLocation();
